Title Find differential abundance within a dataset.

This is a wrapper function that use ALDEx2 package to find the taxa that are differentially abundant between two groups. You have to provide a phyloseq object with only two groups in order to use this function.

differential_abundance(
  physeq,
  group,
  paired = F,
  col1 = "Genus",
  col2 = NULL,
  cores = F,
  plot = F
)

Arguments

physeq

Phyloseq object to compute analysis.

group

The group to perform analysis. Make sure the vector is of length 2. If not please subset your data.

paired

Are the data paired.

col1

First column used to name the taxa in the plot. Default is Genus level.

col2

Second column used to name the taxa in the plot. Default is the last column of your tax_table. It is recommended to use an ASV or OTU number for each taxa.

Value

A list of three elements.

all_features

A dataframe containing all taxa tested with their respective values.

signif_features

A dataframe containing only taxa significant for eBH test.

barplot

Differential abundance plotted witha barplot

volcano

Differential abundance plotted witha volcano plot