Check the dispersion of your samples and the impact of the zero_treatment.

Check the dispersion of your samples and the impact of the zero management you want to perform. There are three possibilities random_generation(), no_zero() and minimum_count(). It will produce a list of plots were you will find the \(neg1/neg2\) ratio and the \(pos/neg1\) ratio. This is particularly important to check once you're working on a new set of data. Take into account that random_generation will be usefull for samples with a few taxa. If you samples are from healthy adults you might want to use

neg_dispersion(
  df,
  positive_sorted_sample = "pos",
  negative_sorted_sample = "neg1",
  second_negative_sample = "neg2",
  type = "facet_all_three"
)

Arguments

df

A element of the seq_table() returned object. You can run this function using a loop or lapply to make the plots for every samples.

positive_sorted_sample

The pattern of the positive fraction. Default = "pos"

negative_sorted_sample

The pattern of the first negative fraction. Default = "neg1"

second_negative_sample

The pattern of the second negative fraction. Default = "neg2"

type

Either a facet of the three possibilities as a single panel: meaning the random_generation(), the minimum() or no_zero() functions as a single panel or superposed : the \(pos/neg1\) ratio for each way of dealing with zero values on one panel and all the \(neg1/neg2\) ratio on an other panel.

Value

Return a ggplot2 graph.

facet_all_three

For the facet_all_three : the black dots will represent the no_zero(), the green dots the random_generation() and the red dots the minimum_count().

superposed

For the superposed : the grey dots represent the \(neg1/neg2\) and the red dots represent \(pos/neg1\).

Examples