A function to plot heatmap nicely, using a phylogenetic distance.

A function to plot heatmap nicely, using a phylogenetic distance.

phylo_heatmap(
  physeq = physeq,
  labels = NULL,
  top = 30,
  factor_to_plot = NULL,
  distance = "wunifrac",
  taxa_rank = "Genus",
  split = NULL,
  color_vector = c("white", "#88CCAA", "#771122"),
  taxa_names_par = gpar(fontsize = 15, fontface = "bold.italic"),
  label_names_par = gpar(fontsize = 15, fontface = "bold")
)

Arguments

physeq

(Required). A phyloseq object.

top

The top number of ASV you want to represent. Default is 30. It's usually the best number to plot.

factor_to_plot

(Optional) Factor for the title.

distance

Phylogenetic dinstance to use as dendrogramm. See beta_diversity.

taxa_rank

The taxa rank at which you want to agregate abundances. Default is "Genus".

split

(Careful !). Numeric number by which you want the dendrogramm to be split. Work only with ComplexHeatmap version 2.2.0.

color_vector

A character string for color plottinh. Default is c("white","#88CCAA","#771122").

taxa_names_par

Graphical parameters. Default is gpar(fontsize= 15, fontface= "bold.italic")

label_names_par

Graphical paramaters. Default is gpar(fontsize=15, fontface="bold")

Examples

library(phyloseq)
data(enterotype)
phylo_heatmap(physeq=enterotype, top=30, labels=NULL, distance="bray",
taxa_rank="Family")
#> [1] "No phylogenetic tree in this phyloseq object, bray-curtis distance selected."
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'taxa' in selecting a method for function 'prune_taxa': error in evaluating the argument 'object' in selecting a method for function 'otu_table': Bad taxrank argument. Must be among the values of rank_names(physeq)