Make a constrained reduction on matrices and data.frame.

Currently the function supports only the constrained reduction around categorical data. It's not possible yet to give two different Omics datasets to find correlation between them.

plot_constrained_reduction(
  mat,
  clinical_data,
  axis_x = 1,
  axis_y = 2,
  model,
  nf = 5,
  method = "CCA",
  type = "boxplot",
  group = NULL,
  stat = "none",
  color_vector = c("cyan4", "brown", "deepskyblue", "black", "red"),
  legend_title = NULL,
  lwd = 1,
  conf = 0.9,
  cex = 2,
  font = 2,
  pch = 20,
  draw = "lines",
  ylimits = "auto",
  xlimits = "auto",
  text = F,
  ncol = 1,
  x.intersp = 1,
  y.intersp = 0.5,
  where = "topleft",
  inset = 0.2,
  stat.cex = 2,
  ...
)

Arguments

mat

A numeric matrix with samples as rows

clinical_data

An integer specifying the columns that are discrete values.

axis_x

Which component for axis x. Default = 1

axis_y

Which component for axis y. Default = 2

model

Specify here the model to be used for regression

nf

Number of components

method

The mathematical method to reduce the dimensions. Currently this function supports CCA, RDA and dbRDA

type

Chose the type of plot you want, choices are : "boxplot", "pure" or "arrows". Default="boxplot".

group

String defining the groups you want to plot on the beta diversity.

stat

Either permanova adonis2 with default parameter or envfit function.

color_vector

Optionnal. A color vector you wish to use for plotting

legend_title

Optionnal. A legend title.

lwd

Line width

conf

Confidence interval. Default = 0.9.

cex

Text size

font

Font type

pch

Shape type

draw

Categorical, draw ellipses as lines or as polygons#'

ylimits

Limits for Y axis, if not specified they will be automatically set

xlimits

Limits for X axis, if not specified they will be automatically set

text

Plot labels or not.

ncol

Number of columns for the legend. Default = 1.

x.intersp

character interspacing factor for horizontal (x) spacing between symbol and legend text.

y.intersp

vertical (y) distances (in lines of text shared above/below each legend entry). A vector with one element for each row of the legend can be used.

where

Where you want the permanova result.

inset

Permanova legend adjustment see legend

stat.cex

Size of the stats text.

...

Value

A plot rendered using the base package and vegan functions :

ordispider

ordiellipse

See also