Plot PLS-DA using mixOmics plsda

A ggplot2 based function to plot PLS-DA based on the object returned by plsda. The function will also plot the features loadings, a biplot then. This function can also handle results of a splsda.

plot_plsda(
  res.plsda,
  comp = c(1, 2),
  color_vector = tol21rainbow,
  shape,
  top.loads = 10
)

Arguments

res.plsda

An object returned by plsda.

comp

Which component to plot.

color_vector

Color vector

shape

Shape of the dots.

top.loads

Top N features to plot. Default = 10. Loads are defined by the absolute value of each features for each components. If top.loads = 10 the function will keep the top10 features of component 1 and the top 10 features of the component 2.

Examples

data(metabolomic)
res= mixOmics::plsda(X = metabolomic[,5:217], Y = metabolomic$birth_type, ncomp=5)

# Basic plotting
plot_plsda(res, comp= c(1,2), shape=21, top.loads= 20)
#> Joining with `by = join_by(rowname, comp1, comp2, comp3, comp4, comp5)`
#> New names:
#> • `` -> `...6`
#> Error in plot_plsda(res, comp = c(1, 2), shape = 21, top.loads = 20): object 'tol21rainbow' not found

# Improve a bit the plot.
plot_plsda(res, comp= c(1,2), shape=21, top.loads= 20, color_vector= tol21rainbow[c(2,14)])+
guides(fill= guide_legend("Birth route"))+
theme_bw()+
theme(legend.position = "bottom",
       axis.title = element_text(size=15, face="bold"),
       panel.grid = element_blank(),
       legend.text = element_text(size=15),
       legend.title = element_text(size=20, face="bold"),
       axis.text = element_text(size=15))
#> Joining with `by = join_by(rowname, comp1, comp2, comp3, comp4, comp5)`
#> New names:
#> • `` -> `...6`
#> Error in eval(expr, envir, enclos): object 'tol21rainbow' not found