Title Screen a model on all your variables.

A function to test machine learning on your different variables of the metadata. This function supports phyloseq object as input or a list containing a values matrix and a variable dataframe. This function is based on caret.

screen_models(
  obj,
  model = "plsda",
  cores = 1,
  ncomp = 10,
  number = 10,
  repeats = 3,
  prop = 0.5,
  train = T,
  ...
)

Arguments

obj

An object of class phyloseq or a list of two levels containing a matrix of values and a dataframe of variables.

model

The model tested, for now this function supports randomForest::randomForest(), glmnet::glmnet() and caret::plsda().

cores

The number of cores you want to use.

ncomp

Number of component you want to keep for caret::plsda(). Default =10.

number

Either the number of folds or number of resampling iterations

repeats

For repeated k-fold cross-validation only: the number of complete sets of folds to compute. Default = 3

prop

The proportion of the variables to use in the training model. Default = 0.5.

train

Wheter or not you want to train the models in order to predict. If you select FALSE the function will use all the data to perform classification or find the best features that explain your data according to the variables. Default=T.

...

Argument passed to caret::train().

Value

A list containing

model

A caret::train() class

Y

The trimmed variables

X

The trimmed values

train_y

The variables used to train the model

train_x

The values used to train the model

pred_x

The predicted values

pred_y

The predicted classification

Examples

screen_models(enterotype, model="glmnet", prop=0.5, ncomp=20)
#> [1] "Sample_ID : is constituted of unique observations"
#> Warning: Some classes have a single record ( AM.AD.1, AM.AD.2, AM.F10.T1, AM.F10.T2, DA.AD.1, DA.AD.2, DA.AD.3, DA.AD.4, ES.AD.1, ES.AD.2, ES.AD.3, ES.AD.4, FR.AD.1, FR.AD.2, FR.AD.3, FR.AD.4, FR.AD.5, FR.AD.6, FR.AD.7, FR.AD.8, IT.AD.1, IT.AD.2, IT.AD.3, IT.AD.4, IT.AD.5, IT.AD.6, JP.AD.1, JP.AD.2, JP.AD.3, JP.AD.4, JP.AD.5, JP.AD.6, JP.AD.7, JP.AD.8, JP.AD.9, JP.IN.1, JP.IN.2, JP.IN.3, JP.IN.4 ) and these will be selected for the sample
#> Warning: Some classes have a single record ( CD, UC ) and these will be selected for the sample
#> Error in trainControl(method = "repeatedcv", number = number, repeats = repeats): could not find function "trainControl"

screen_models(enterotype, model="plsda", cores=1, ncomp=10, number=10, repats=5, prop=0.6, train=T)
#> [1] "Sample_ID : is constituted of unique observations"
#> Warning: Some classes have a single record ( AM.AD.1, AM.AD.2, AM.F10.T1, AM.F10.T2, DA.AD.1, DA.AD.2, DA.AD.3, DA.AD.4, ES.AD.1, ES.AD.2, ES.AD.3, ES.AD.4, FR.AD.1, FR.AD.2, FR.AD.3, FR.AD.4, FR.AD.5, FR.AD.6, FR.AD.7, FR.AD.8, IT.AD.1, IT.AD.2, IT.AD.3, IT.AD.4, IT.AD.5, IT.AD.6, JP.AD.1, JP.AD.2, JP.AD.3, JP.AD.4, JP.AD.5, JP.AD.6, JP.AD.7, JP.AD.8, JP.AD.9, JP.IN.1, JP.IN.2, JP.IN.3, JP.IN.4 ) and these will be selected for the sample
#> Warning: Some classes have a single record ( CD, UC ) and these will be selected for the sample
#> Error in screen_models(enterotype, model = "plsda", cores = 1, ncomp = 10,     number = 10, repats = 5, prop = 0.6, train = T): object 'train_y' not found