Analyse the ratio between your IgA positive fraction and the two IgA negative fractions.

The slide_z function will make a Z test based on the log2 ratio of \(pos*neg1\) (slide_z_standard) and \(neg1/neg2\) (slid_z_modern).

slide_z(
  df,
  positive_sorted_sample = "pos",
  negative_sorted_sample = "neg1",
  second_negative_sample = "neg2",
  zero_treatment = "random generation",
  deltaX = 50,
  alpha = 0.05,
  slide_version = "slide_z_modern",
  confidence_interval = c(0.95, 0.99, 0.999),
  plot = FALSE
)

Arguments

df

(Required). An object of class dataframe or matrix generated using seq_table

positive_sorted_sample

Character present in every positive sorted samples, for example "pos"

negative_sorted_sample

Character present in every negative sorted samples, for example "neg1"

second_negative_sample

Character present in every negative sorted samples, for example "neg2"

zero_treatment

(Default = "random generation") The treatment you wish to apply to ASV equal to zero in one or two samples in order to compare it with a non zero sample. Random generation will generate random numbers between the minimum value different from zero and the 5\ data.

deltaX

(Default = 50) The number of ASV to take at a time to create the main algorithm. See Details for more information

alpha

The alpha

slide_version

(Soon to be defunct) slide_z_modern or slide_z_standard. The standard version (not recommended) will use the log2(pos/neg1) ratio to create the normal distribution and deviation while slide_z_modern will use the log2(neg1/neg2) ratio to perform the same step.

confidence_interval

(Default= c("0.95", "0.99", "0.999")) The confidence interval for the ellipse to return.

plot

(Default= FALSE) Wether you want to print the slide_z plots. If TRUE, a plot of log2(pos/neg1) and log10(pos*neg1) with confidence ellipse will be drawn using ggplot2 and plotly with the taxonomy as tool. Plotting will slow drastically the code.

Value

Return an IgaSeq object containing : The function will return a single IgASeq for each samples in this present version of the package. You need to perform a for loop or lapply to make an IgASeq for each samples. The IgASeq object is composed of the following :

ig_seq_all

All the taxa found in the phyloseq object.

ig_up

All the taxa that are enriched in the positive fraction

ig_down

All the taxa that are enriched in the negative fraction

Additionally, each slot contains :

score

The IgASeq score obtained with the slide_z

ellipse

The value of confidence for a given taxa.

log_ratio values

Refers to log_ratio for more information on the values given by log_ratio.